test(cypress): abstract data in function
