+++ packages/kit-headless/src/components/autocomplete/test-data.ts
@@ -1,0 +1,85 @@
+export default function TestData() {
+  const fruits = [
+    'Apple',
+    'Apricot',
+    'Avocado 🥑',
+    'Banana',
+    'Bilberry',
+    'Blackberry',
+    'Blackcurrant',
+    'Blueberry',
+    'Boysenberry',
+    'Currant',
+    'Cherry',
+    'Coconut',
+    'Cranberry',
+    'Cucumber',
+    'Custard apple',
+    'Damson',
+    'Date',
+    'Dragonfruit',
+    'Durian',
+    'Elderberry',
+    'Feijoa',
+    'Fig',
+    'Gooseberry',
+    'Grape',
+    'Raisin',
+    'Grapefruit',
+    'Guava',
+    'Honeyberry',
+    'Huckleberry',
+    'Jabuticaba',
+    'Jackfruit',
+    'Jambul',
+    'Juniper berry',
+    'Kiwifruit',
+    'Kumquat',
+    'Lemon',
+    'Lime',
+    'Loquat',
+    'Longan',
+    'Lychee',
+    'Mango',
+    'Mangosteen',
+    'Marionberry',
+    'Melon',
+    'Cantaloupe',
+    'Honeydew',
+    'Watermelon',
+    'Miracle fruit',
+    'Mulberry',
+    'Nectarine',
+    'Nance',
+    'Olive',
+    'Orange',
+    'Clementine',
+    'Mandarine',
+    'Tangerine',
+    'Papaya',
+    'Passionfruit',
+    'Peach',
+    'Pear',
+    'Persimmon',
+    'Plantain',
+    'Plum',
+    'Pineapple',
+    'Pomegranate',
+    'Pomelo',
+    'Quince',
+    'Raspberry',
+    'Salmonberry',
+    'Rambutan',
+    'Redcurrant',
+    'Salak',
+    'Satsuma',
+    'Soursop',
+    'Star fruit',
+    'Strawberry',
+    'Tamarillo',
+    'Tamarind',
+    'Yuzu',
+  ];
+
+  return fruits;
+}
